Add basic support for visit types
Up until now OCEMR was designed to handle only outpatient visits.
Everyone would check out by the end of the day. This feature adds
support for generic inpatients and maternity patient types. the main
patient queue can display all patients or just focus on a single type of
patient that the clinician is interested in.